requirements.txt
PyQt5>=5.15.0
Pillow>=10.0.0
pypdf>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
olefile>=0.46
xlrd>=2.0.0
xlwt>=1.3.0
pywin32>=300; sys.platform == 'win32'
